prov/efa: Split efa_rdm_domain and efa_domain into separate files
This patch separates the previously split domain structure into different
files. The base efa_domain stays in prov/efa/src/efa_domain.{c,h};
the wrapper efa_rdm_domain moves to prov/efa/src/rdm/efa_rdm_domain.{c,h}.
Domain unit tests are also split.

Renames and modifies the input arguments for helpers that are only used
in the rdm domain path. Renames variables as needed to maintain clarity
between base and extended structures.
