+++ prov/efa/src/efa_av.c
@@ -11,6 +11,7 @@
 
 #include "efa.h"
 #include "efa_av.h"
+#include "rdm/efa_rdm_domain.h"
 #include "rdm/efa_rdm_pke_utils.h"
 
 static inline struct efa_conn *efa_av_addr_to_conn_impl(struct util_av *util_av,
