+++ prov/efa/src/rdm/efa_rdm_cntr.c
@@ -106,15 +106,13 @@ static void efa_rdm_cntr_progress(struct util_cntr *cntr)
 {
 	struct dlist_entry *item;
 	struct efa_rdm_cntr *efa_rdm_cntr;
-	struct efa_domain *efa_domain;
 	struct efa_rdm_domain *rdm_domain;
 	struct efa_rdm_ep *efa_rdm_ep;
 	struct fid_list_entry *fid_entry;
 
 	ofi_genlock_lock(&cntr->ep_list_lock);
 	efa_rdm_cntr = container_of(cntr, struct efa_rdm_cntr, efa_cntr.util_cntr);
-	efa_domain = container_of(efa_rdm_cntr->efa_cntr.util_cntr.domain, struct efa_domain, util_domain);
-	rdm_domain = (struct efa_rdm_domain *) efa_domain;
+	rdm_domain = (struct efa_rdm_domain *) container_of(efa_rdm_cntr->efa_cntr.util_cntr.domain, struct efa_domain, util_domain);
 
 	/**
 	 * TODO: It's better to just post the initial batch of internal rx pkts during ep enable
@@ -134,7 +132,7 @@ static void efa_rdm_cntr_progress(struct util_cntr *cntr)
 	}
 
 	efa_cntr_progress_ibv_cq_poll_list(&efa_rdm_cntr->efa_cntr);
-	efa_domain_progress_rdm_peers_and_queues(rdm_domain);
+	efa_rdm_domain_progress_peers_and_queues(rdm_domain);
 	ofi_genlock_unlock(&cntr->ep_list_lock);
 }
 
@@ -143,7 +141,6 @@ int efa_rdm_cntr_open(struct fid_domain *domain, struct fi_cntr_attr *attr,
 {
 	int ret;
 	struct efa_rdm_cntr *cntr;
-	struct efa_domain *efa_domain;
 	struct efa_rdm_domain *rdm_domain;
 	struct fi_cntr_attr shm_cntr_attr = {0};
 	struct fi_peer_cntr_context peer_cntr_context = {0};
@@ -153,9 +150,8 @@ int efa_rdm_cntr_open(struct fid_domain *domain, struct fi_cntr_attr *attr,
 		return -FI_ENOMEM;
 
 	cntr->need_to_scan_ep_list = false;
-	efa_domain = container_of(domain, struct efa_domain,
-				  util_domain.domain_fid);
-	rdm_domain = (struct efa_rdm_domain *) efa_domain;
+	rdm_domain = (struct efa_rdm_domain *) container_of(domain, struct efa_domain,
+							    util_domain.domain_fid);
 
 	ret = efa_cntr_construct(&cntr->efa_cntr, domain, attr,
 				 efa_rdm_cntr_progress, context);
