+++ prov/efa/src/rdm/efa_rdm_ep_utils.c
@@ -185,9 +185,9 @@ struct efa_rdm_ope *efa_rdm_ep_alloc_rxe(struct efa_rdm_ep *ep, struct efa_rdm_p
 	}
 
 	rxe->ep = ep;
-	efa_domain_ope_list_lock(efa_rdm_ep_rdm_domain(ep));
+	efa_rdm_domain_ope_list_lock(efa_rdm_ep_rdm_domain(ep));
 	dlist_insert_tail(&rxe->ep_entry, &ep->rxe_list);
-	efa_domain_ope_list_unlock(efa_rdm_ep_rdm_domain(ep));
+	efa_rdm_domain_ope_list_unlock(efa_rdm_ep_rdm_domain(ep));
 	rxe->type = EFA_RDM_RXE;
 	rxe->internal_flags = 0;
 	rxe->fi_flags = 0;
