+++ prov/efa/test/efa_unit_tests.h
@@ -358,31 +358,39 @@ void test_efa_rdm_pke_get_available_copy_methods_align128(void **state);
 /* begin efa_unit_test_domain.c */
 void test_efa_domain_info_type_efa_direct(void **state);
 void test_efa_domain_direct_has_bounce_buffer(void **state);
-void test_efa_domain_rdm_no_bounce_buffer(void **state);
 void test_efa_domain_no_bounce_buffer_without_fi_rma_cap_requested(void **state);
 void test_efa_domain_bounce_buffer_with_rdma(void **state);
-void test_efa_domain_info_type_efa_rdm(void **state);
 void test_efa_domain_open_ops_wrong_name(void **state);
 void test_efa_domain_open_ops_mr_query(void **state);
-void test_efa_domain_rdm_attr_mr_allocated(void **state);
 void test_efa_domain_dgram_attr_mr_allocated(void **state);
 void test_efa_domain_direct_attr_mr_allocated(void **state);
-void test_efa_domain_peer_list_cleared(void **state);
 void test_efa_domain_open_ops_query_addr(void **state);
 void test_efa_domain_open_ops_query_qp_wqs(void **state);
 void test_efa_domain_open_ops_query_cq(void **state);
 void test_efa_domain_open_ops_cq_open_ext(void **state);
+void test_efa_domain_open_ops_cntr_open_ext(void **state);
 void test_efa_domain_open_ops_get_mr_lkey(void **state);
 void test_efa_fabric_open_ops_feature_known(void **state);
 void test_efa_fabric_open_ops_feature_not_on_proto(void **state);
 void test_efa_fabric_open_ops_feature_unknown(void **state);
-void test_efa_domain_rdm_mr_ops(void **state);
 void test_efa_domain_direct_mr_ops(void **state);
 void test_efa_domain_dgram_mr_ops(void **state);
+void test_efa_domain_open_installs_base_domain_ops_efa_direct(void **state);
+void test_efa_domain_open_installs_base_domain_ops_dgram(void **state);
+void test_efa_domain_gda_ops_rejected_for_dgram(void **state);
+/* end efa_unit_test_domain.c */
+
+/* begin efa_unit_test_rdm_domain.c */
+void test_efa_domain_info_type_efa_rdm(void **state);
+void test_efa_domain_rdm_no_bounce_buffer(void **state);
+void test_efa_domain_rdm_attr_mr_allocated(void **state);
+void test_efa_domain_peer_list_cleared(void **state);
+void test_efa_domain_rdm_mr_ops(void **state);
 void test_efa_domain_mr_cache_enabled(void **state);
 void test_efa_domain_mr_cache_disabled_with_mr_local(void **state);
-void test_efa_domain_open_ops_cntr_open_ext(void **state);
-/* end efa_unit_test_domain.c */
+void test_efa_rdm_domain_open_installs_rdm_domain_ops(void **state);
+void test_efa_domain_gda_ops_rejected_for_rdm(void **state);
+/* end efa_unit_test_rdm_domain.c */
 
 void test_efa_rdm_cq_ibv_cq_poll_list_same_tx_rx_cq_single_ep(void **state);
 void test_efa_rdm_cq_ibv_cq_poll_list_separate_tx_rx_cq_single_ep(void **state);
